requirements.txt
PyCap[data_science] == 2.6.0
python-dotenv == 1.0.0
